% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocSwirl-package.R
\docType{package}
\name{BiocSwirl-package}
\alias{BiocSwirl}
\alias{BiocSwirl-package}
\title{BiocSwirl: Interactive Bioinformatics Tutorial on scRNA-Seq}
\description{
BiocSwirl aims to make learning bioinformatics concepts hands on through the development of course material that takes you through the common bioinformatics workflows (with a focus on scRNA-seq).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lisancao/biocswirl}
}

}
\author{
\strong{Maintainer}: Lisa Cao \email{lisanatashacao@gmail.com}

Authors:
\itemize{
  \item Mariam Arab \email{arab@sfu.ca}
  \item Kate Tyshchenko \email{kate.tyshchenko@yandex.ru}
  \item Jackie Lu \email{jql6@sfu.ca}
  \item Jeremy Fan \email{jeremy.fan@ucalgary.ca}
  \item Paaksum Wong \email{paaksumwong@gmail.com}
  \item Sourav Singh \email{ssouravsingh12@gmail.com}
}

}
\keyword{internal}
