% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_course.R
\name{install_course}
\alias{install_course}
\title{Install and open a course of choice}
\usage{
install_course(course)
}
\arguments{
\item{course}{a course name, input as a string}
}
\description{
This function installs and opens a course selected by the user.
}
\examples{
install_course("scrna-seq")
}
\keyword{install}
\keyword{open}
