% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_courses.R
\name{installed_courses}
\alias{installed_courses}
\title{This function will list all the installed BiocSwirl courses}
\usage{
installed_courses()
}
\description{
This function will list all the installed BiocSwirl courses
}
\examples{
\dontrun{
installed_courses()
}
}
