% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{list_courses}
\alias{list_courses}
\title{list courses
This function will list all available BiocSwirl courses available for installation}
\usage{
list_courses()
}
\description{
list courses
This function will list all available BiocSwirl courses available for installation
}
