% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{load_all}
\alias{load_all}
\title{This function will load and install all available BiocSwirl courses for you}
\usage{
load_all()
}
\description{
This function will load and install all available BiocSwirl courses for you
}
\examples{
load_all()
}
