% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{load_course}
\alias{load_course}
\title{A function to load and install a specificed BiocSwirl course for you}
\usage{
load_course(course)
}
\arguments{
\item{course}{options are 'RNAseq' and 'Intro_to_Data_Science'. More information on available courses via list_courses()}
}
\description{
A function to load and install a specificed BiocSwirl course for you
}
\examples{
\dontrun{
load_course('RNAseq')
load_course('Intro_to_Data_Science')
}
}
