% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{load_courses_warning}
\alias{load_courses_warning}
\title{This function will load and install all available BiocSwirl courses for you}
\usage{
load_courses_warning(course)
}
\arguments{
\item{course}{name of the course that could produce a warning like 'mbb343' if already installed}
}
\description{
and create warnings (for demo purposes) because I was too scared to include this in the other function.
}
\examples{
load_courses_warning('mbb343')
}
