% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{start_course}
\alias{start_course}
\title{This function will start the course environment for you}
\usage{
start_course()
}
\description{
This function will start the course environment for you
}
\examples{
\dontrun{
start_course()
}
}
