% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_courses.R
\name{uninstall_all_courses}
\alias{uninstall_all_courses}
\title{This function will uninstall all the installed BiocSwirl courses}
\usage{
uninstall_all_courses()
}
\description{
This function will uninstall all the installed BiocSwirl courses
}
\examples{
\dontrun{
uninstall_all_courses()
}
}
