% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_courses.R
\name{uninstall_course}
\alias{uninstall_course}
\title{This function will uninstall a selected course}
\usage{
uninstall_course(course)
}
\arguments{
\item{course}{string identifying one of the installed BiocSwirl courses that should be uninstalled. check installed_courses().}
}
\description{
This function will uninstall a selected course
}
\examples{
\dontrun{
uninstall_course('BiocSwirl_RNAseq')
}
}
