% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_courses.R
\name{list_courses}
\alias{list_courses}
\title{This function will list all available BiocSwirl courses available for installation}
\usage{
list_courses()
}
\description{
This function will list all available BiocSwirl courses available for installation
}
\examples{
\dontrun{
list_courses()
}
}
