% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_courses.R
\name{list_courses}
\alias{list_courses}
\title{Course list function}
\usage{
list_courses()
}
\description{
This function lists all the course titles available.
}
\examples{
list_courses()
}
\keyword{course_list}
